+++ src/library/DIPS.Mobile.UI/Components/TabView/Android/TabViewHandler.cs
@@ -25,6 +25,8 @@ public class TabViewHandler : ViewHandler<TabView, Google.Android.Material.Tabs.
         var tabLayout = new TabLayout(Context);
         tabLayout.TabMode = TabLayout.ModeScrollable;
         tabLayout.SetBackgroundColor(Colors.GetColor(ColorName.color_palette_neutral_transparent).ToPlatform());
+        var density = Context.Resources.DisplayMetrics.Density;
+        tabLayout.SetPadding(0, 0, 0, (int)(Sizes.GetSize(SizeName.size_2) * density));
         m_tabLayout = tabLayout;
         
         return tabLayout;
@@ -45,14 +47,40 @@ public class TabViewHandler : ViewHandler<TabView, Google.Android.Material.Tabs.
         platformView.TabSelected -= OnTabSelected;
         base.DisconnectHandler(platformView);
     }
-
-    void OnTabSelected(object sender, TabLayout.TabSelectedEventArgs e)
+    
+    async void OnTabSelected(object sender, TabLayout.TabSelectedEventArgs e)
     {
+        if (await TrySwitchTab())
+        {
+            return;
+        }
+
         var index = e.Tab.Position;
         if (VirtualView.ItemsSource != null && index >= 0 && index < VirtualView.ItemsSource.Count)
         {
-            VirtualView.SelectedTabIndex = index;
+            VirtualView.SetSelectedTabIndex(index);
         }
+    }
+
+    // This method is called when the user tries to switch tabs, to check if a confirmation is needed and if switching shall proceed.
+    private async Task<bool> TrySwitchTab()
+    {
+        if (VirtualView.CanSwitchTab != null)
+        {
+            UpdateTabColor();
+            var result = await VirtualView.CanSwitchTab(VirtualView.SelectedTabIndex);
+            if (!result)
+            {
+                m_tabLayout.TabSelected -= OnTabSelected; // Temporarily remove the event handler to prevent recursion
+                m_tabLayout.SelectTab(m_tabLayout.GetTabAt(VirtualView.SelectedTabIndex));
+                UpdateTabColor();
+                m_tabLayout.TabSelected += OnTabSelected;
+
+                return true;
+            }
+        }
+
+        return false;
     }
 
     public static void MapItemsSource(TabViewHandler handler, TabView tabView)